Move fragment to new file
